Move all transform tests to SceneDataTransformer
