+++ src/querying/SceneDataTransformer.ts
